+++ src/site/markdown/changelog.md
@@ -6,6 +6,7 @@ The changelog should give you a quick overview of what change between the differ
 
 ### New Features
 
+- [6](https://github.com/FHOOEAIST/imaging/issues/6) - Added support for ImageJ based images.
 - [11](https://github.com/FHOOEAIST/imaging/issues/11) - New `TypeBasedImageFactoryFactory`, that allows to create an image factory for a given type.
 
 ### Dependencies